move node, schemas and serialization tests
