+++ core-tests/src/test/java/net/corda/coretests/schemas/GoodSchemaJavaV1.java
@@ -1,4 +1,7 @@
-package net.corda.core.schemas;
+package net.corda.coretests.schemas;
+
+import net.corda.core.schemas.MappedSchema;
+import net.corda.core.schemas.PersistentState;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
