+++ core-tests/src/test/java/net/corda/coretests/schemas/TestJavaSchemaFamily.java
@@ -1,4 +1,4 @@
-package net.corda.core.schemas;
+package net.corda.coretests.schemas;
 
 public class TestJavaSchemaFamily {
 }
